requirements.txt
pandas==2.0.3
numpy==1.25.2
requests==2.32.3
streamlit==1.40.0
sseclient-py==1.8.0
pydantic==2.7.3
urllib3 >= 2.1.0, < 3.0.0
python_dateutil >= 2.8.2
pydantic >= 2
typing-extensions >= 4.7.1